# Only cache transformed work-free top-level up to a certain size