--- a/clash-lib/src/Clash/Normalize.hs
+++ b/clash-lib/src/Clash/Normalize.hs
@@ -144,6 +144,7 @@ runNormalization opts supply globals typeTrans reprs tcm tupTcm eval primMap rcs
                   rcsMap
                   (opt_newInlineStrat opts)
                   (opt_ultra opts)
+                  (opt_inlineWFCacheLimit opts)
 
 
 normalize