--- a/clash-lib/src/Clash/Normalize/Types.hs
+++ b/clash-lib/src/Clash/Normalize/Types.hs
@@ -68,6 +68,8 @@ data NormalizeState
   -- ^ High-effort normalization session, trading performance improvement for
   -- potentially much longer compile times. Follows the 'Clash.Driver.opt_ultra'
   -- flag.
+  , _inlineWFCacheLimit :: !Word
+  -- ^ At what size do we cache normalized work-free top-level binders.
   }
 
 makeLenses ''NormalizeState