And don't apply the full set of normalization transformations on them, only "local" (i.e. not touch other global binders) constant propagation.

This can decrease normalization time on certain tests up to 10 times, such as AES.hs

before:
```
benchmarking normalization of tests/shouldwork/Basic/AES.hs
time                 6.516 s    (6.483 s .. 6.553 s)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 6.493 s    (6.477 s .. 6.505 s)
std dev              15.76 ms   (7.314 ms .. 20.13 ms)
variance introduced by outliers: 19% (moderately inflated)
```

after:
```
benchmarking normalization of tests/shouldwork/Basic/AES.hs
time                 542.5 ms   (504.8 ms .. 591.1 ms)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 526.8 ms   (517.7 ms .. 534.7 ms)
std dev              9.958 ms   (4.435 ms .. 13.13 ms)
variance introduced by outliers: 19% (moderately inflated)
```